+++ templates/spec_helper_acceptance.rb.erb
@@ -1,1 +1,1 @@
-require 'controlrepo/beaker/spec_helper'
+require 'onceover/beaker/spec_helper'
